ei_draw.c

+++ projet_c_ig.3.0.2/projet_c_ig/implem/ei_draw.c
@@ -7,11 +7,12 @@ void ei_fill(ei_surface_t surface, const ei_color_t* color, const ei_rect_t* cli
     hw_surface_lock(surface); 
 
     //Returns a pointer to the address of the pixel at coordinated (0, 0) of the surface.
-    uint32_t* surface_buffer = hw_surface_get_buffer(surface); 
+    uint8_t* surface_buffer = hw_surface_get_buffer(surface); 
 
     //test if color is NULL
     if (color == NULL){
         //painted black (opaque).
+        
 
     } else {
         //paint of the surface with color (don't take clipper into account for now)
